+++ src/PostFinanceCheckoutPayment.php
@@ -3,6 +3,7 @@
 namespace PostFinanceCheckoutPayment;
 
 use Shopware\Core\{
+  	Framework\Feature,
 	Framework\Plugin,
 	Framework\Plugin\Context\ActivateContext,
 	Framework\Plugin\Context\DeactivateContext,
@@ -21,6 +22,7 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\DirectoryLoader;
 use Symfony\Component\DependencyInjection\Loader\GlobFileLoader;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 
 // expect the vendor folder on Shopware store releases
 if (file_exists(dirname(__DIR__) . '/vendor/autoload.php')) {
@@ -82,19 +84,21 @@ class PostFinanceCheckoutPayment extends Plugin {
 	{
 		parent::build($container);
 
-		$locator = new FileLocator('Resources/config');
+		$confDir = \rtrim($this->getPath(), '/') . '/Resources/config';
+		$locator = new FileLocator($confDir);
 
 		$resolver = new LoaderResolver([
-			new YamlFileLoader($container, $locator),
-			new GlobFileLoader($container, $locator),
-			new DirectoryLoader($container, $locator),
+		  new YamlFileLoader($container, $locator),
+		  new XmlFileLoader($container, $locator),
+		  new GlobFileLoader($container, $locator),
+		  new DirectoryLoader($container, $locator),
 		]);
 
 		$configLoader = new DelegatingLoader($resolver);
 
-		$confDir = \rtrim($this->getPath(), '/') . '/Resources/config';
-
 		$configLoader->load($confDir . '/{packages}/*.yaml', 'glob');
+
+		$configLoader->load('services/core/checkout.xml');
 	}
 
 	public function enrichPrivileges(): array